Widen the test fixture to the 65816's 24-bit address space
read6502() and write6502() ignored the bank byte and memory was a flat 64K, so
no 65816 test could tell one bank from another. Memory is now 16 MB indexed by
(bank << 16) | offset, which leaves bank 0 as the first 64K and every existing
test unchanged.

Two things this exposed, both fixed here:

Clearing 16 MB per reset would dominate the run; ProcessorTests resets once per
case, tens of thousands of times. Bank 0 is still cleared outright and pages
above it are cleared only when written. A test that dirties more than 256 pages
falls back to a full clear, which is what the Klaus images do twice.

Indexing that carries past $FFFF now reaches bank 1, where before the truncation
to uint16_t wrapped it. A 65C02 has sixteen address lines and no bank pins, so
the fixture forces bank 0 unless the core is running as a 65816. Whether the
emulator's own memory path has the same exposure is recorded separately.

Klaus read its images with sizeof cpu_mem, which is no longer 64K.

No behaviour change: state stays 25400/25400 and cycles 25300/25400, all 15
tests pass, 11/11 mutations caught.

+++ tests/test_processor_tests.c
@@ -53,10 +53,10 @@ static pt_cycle bus[BUS_MAX];
 static int      bus_n;
 
 static void
-note_read(uint16_t addr, uint8_t value)
+note_read(uint32_t addr, uint8_t value)
 {
 	if (bus_n < BUS_MAX) {
-		bus[bus_n].addr = addr;
+		bus[bus_n].addr = (uint16_t)addr;
 		bus[bus_n].value = value;
 		bus[bus_n].is_write = 0;
 		bus_n++;
@@ -64,10 +64,10 @@ note_read(uint16_t addr, uint8_t value)
 }
 
 static void
-note_write(uint16_t addr, uint8_t value)
+note_write(uint32_t addr, uint8_t value)
 {
 	if (bus_n < BUS_MAX) {
-		bus[bus_n].addr = addr;
+		bus[bus_n].addr = (uint16_t)addr;
 		bus[bus_n].value = value;
 		bus[bus_n].is_write = 1;
 		bus_n++;
